requirements.txt
functions-framework==3.*
pandas==1.5.3
sqlalchemy==2.0.19
psycopg2==2.9.9
python-dotenv==1.0.0
google-cloud-storage==1.44.0